clean up css, rename modal, still need to change shape of generate network and cancel buttons
