+++ web-client/package.json
@@ -27,7 +27,6 @@
     "@vitejs/plugin-react": "^5.0.0",
     "@vitest/coverage-v8": "^4.0.10",
     "globals": "^16.3.0",
-    "parallelshell": "^3.0.1",
     "prettier": "3.6.2",
     "vite": "^7.1.2",
     "vitest": "^4.0.10"
